Refactor departures queries, show journey from timetable click.
