+++ src/components/sidepanel/Timetables.js
@@ -1,7 +1,7 @@
 import React, {Component} from "react";
 import {observer} from "mobx-react";
 import SidepanelList from "./SidepanelList";
-import StopTimetable from "../map/StopTimetable";
+import StopTimetable from "./StopTimetable";
 import {Text} from "../../helpers/text";
 import withStop from "../../hoc/withStop";
