+++ src/hoc/withSelectedJourney.js
@@ -9,18 +9,12 @@ export default (Component) => {
   @inject(app("state"))
   @withHfpData
   @observer
-  class WithSelectedJourneyHfp extends React.Component {
+  class WithSelectedJourneyComponent extends React.Component {
     render() {
       const {state, positions} = this.props;
 
       if (!state.selectedJourney) {
-        return (
-          <Component
-            key="withSelectedJourneyComponent"
-            {...this.props}
-            selectedJourneyHfp={[]}
-          />
-        );
+        return <Component {...this.props} selectedJourneyHfp={[]} />;
       }
 
       const selectedJourneyId = getJourneyId(state.selectedJourney);
@@ -31,7 +25,6 @@ export default (Component) => {
 
       return (
         <Component
-          key="withSelectedJourneyComponent"
           {...this.props}
           selectedJourneyHfp={get(journeyHfp, "positions", [])}
         />
@@ -39,5 +32,5 @@ export default (Component) => {
     }
   }
 
-  return WithSelectedJourneyHfp;
+  return WithSelectedJourneyComponent;
 };
